Refactor: migrate diagnostics to use trait rather than macro
This commit makes it so that diagnostic checkers no longer have to
define magic public variables which will be called by a macro. Instead,
they all implement a common trait which will easily enforce that all the
checking data is defined. This also means compile errors will be quicker and
more understandable.